# shared/schemas/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Auto-Order API",
    "version": "1.0.0",
    "description": "API documentation for Auto-Order backend - Personal Purchase Tracker with CRUD operations for Products and Orders, plus purchase prediction capabilities."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Google ID Token as Bearer token"
      }
    },
    "schemas": {
...
      "OrderLineItem": {
        "type": "object",
        "required": [
          "productId",
          "quantity"
        ],
        "properties": {
          "productId": {
            "type": "string",
            "description": "ID of the product"
          },
          "quantity": {
            "type": "number",
            "description": "Quantity of the product"
          },
          "productName": {
            "type": "string",
            "description": "Optional name of the product (for on-the-fly creation)"
          }
        },
        "example": {
          "productId": "123e4567-e89b-12d3-a456-426614174000",
          "quantity": 2,
          "productName": "Coffee Beans"
        }
      },
      "Order": {
        "type": "object",
        "required": [
          "name",
          "creationDate",
          "purchaseDate",
          "lineItems"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the order"
          },
          "name": {
            "type": "string",
            "description": "Name of the order"
          },
          "creationDate": {
            "type": "string",
            "format": "date",
            "description": "Date when the order was created (ISO format)"
          },
          "purchaseDate": {
            "type": "string",
            "format": "date",
            "description": "Date when the order was/will be purchased (ISO format)"
          },
          "lineItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderLineItem"
            },
            "description": "List of items in the order"
          }
        },
        "example": {
          "id": "456e7890-e89b-12d3-a456-426614174001",
          "name": "Weekly Grocery Order",
          "creationDate": "2025-06-19",
          "purchaseDate": "2025-06-20",
          "lineItems": [
            {
              "productId": "123e4567-e89b-12d3-a456-426614174000",
              "quantity": 2,
              "productName": "Coffee Beans"
            }
          ]
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/orders": {
      "get": {
        "summary": "Get all orders",
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "includeDrafts",
            "schema": {
              "type": "boolean",
              "default": true
            },
            "description": "Whether to include draft orders"
          },
          {
            "in": "query",
            "name": "draftsOnly",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Whether to return only draft orders"
          }
        ],
        "responses": {
          "200": {
            "description": "List of all orders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new order",
        "tags": [
          "Orders"
        ],
        "description": "Create a new order. If a product doesn't exist in lineItems and productName is provided, it will be created automatically.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "creationDate",
                  "purchaseDate",
                  "lineItems"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "creationDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "purchaseDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "lineItems": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/OrderLineItem"
                    }
                  }
                },
                "example": {
                  "name": "Weekly Grocery Order",
                  "creationDate": "2025-06-19",
                  "purchaseDate": "2025-06-20",
                  "lineItems": [
                    {
                      "productId": "123e4567-e89b-12d3-a456-426614174000",
                      "quantity": 2
                    },
                    {
                      "productName": "New Product",
                      "quantity": 1
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Invalid order data or product not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "array"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{id}": {
      "put": {
        "summary": "Update an order",
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Order ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "creationDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "purchaseDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "lineItems": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/OrderLineItem"
                    }
                  }
                },
                "example": {
                  "name": "Updated Weekly Grocery Order",
                  "creationDate": "2025-06-19",
                  "purchaseDate": "2025-06-21",
                  "lineItems": [
                    {
                      "productId": "123e4567-e89b-12d3-a456-426614174000",
                      "quantity": 3
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "404": {
            "description": "Order not found"
          }
        }
      },
      "delete": {
        "summary": "Delete an order",
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Order ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Order deleted successfully"
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    },
...
  },
...
}